chore: moved query page. added logs to /database section
•
