+++ studio/components/interfaces/Settings/Logs/LogsPreviewer.tsx
@@ -129,6 +129,8 @@ export const LogsPreviewer: React.FC<Props> = ({ projectRef, queryType, override
     })
   }
 
+  console.log('queryType in LogsPreviewr', queryType)
+
   return (
     <div className="h-full flex flex-col flex-grow">
       <PreviewFilterPanel
